GEODE-9046: Refactor null redis structures to separate classes
	* NULL_REDIS_STRING, NULL_REDIS_HASH, NULL_REDIS_SET, and NULL_REDIS_DATA are moved to a separate class
	* This is done so that super classes do not reference subclasses to avoid class loading deadlocks.
