+++ geode-redis/src/main/java/org/apache/geode/redis/internal/data/RedisSet.java
@@ -47,7 +47,6 @@ import org.apache.geode.redis.internal.delta.RemsDeltaInfo;
 
 public class RedisSet extends AbstractRedisData {
 
-  public static final NullRedisSet NULL_REDIS_SET = new NullRedisSet();
   private HashSet<ByteArrayWrapper> members;
 
   @SuppressWarnings("unchecked")
